Revert "* Temporary change with some debug output to see if my suspicion is right."
This reverts commit de004b4e6af578acd17656c9a6f3f6a69fe2a222.

I think I found the issue. It's not the sorting between the proof attempts, but
between the proof items. They don't know about steps, and the offending entries
have all the same value (max time, total time, rule, severity, and location), so
unstable sorting is to be expected. I guess, to stabilize that we need to create
unique ids to make these guys different.